% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_predictions.R
\name{create_nfl_model}
\alias{create_nfl_model}
\title{Create the Model}
\usage{
create_nfl_model(year, dataset, seed = 123)
}
\arguments{
\item{year:}{the year in which the model will test, meaning the model will be based on data prior to that season.}

\item{dataset:}{this should be set to the dataset that is outputted by the create model function.}

\item{seed:}{seed for the model. pick a random number.}
}
\value{
trained glm model
}
\description{
Create the Model
}
