% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_nfl_model_dataset.R
\name{create_nfl_modeldataset}
\alias{create_nfl_modeldataset}
\title{Create the Dataset}
\usage{
create_nfl_modeldataset(keep_latest_performance = FALSE)
}
\arguments{
\item{keep_latest_performance:}{This will determine whether the dataset that is outputted will be the team's latest performance or the full model dataset}
}
\value{
Model Dataset
}
\description{
Create the Dataset
}
