% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_predictions.R
\name{generate_nfl_predictions}
\alias{generate_nfl_predictions}
\title{Predict the NFL}
\usage{
generate_nfl_predictions(years, seed = 123)
}
\arguments{
\item{years:}{the years for which you want to get predictions from nfl games}

\item{seed:}{seed for the model. pick a random number.}
}
\value{
datatable with predictions for games
}
\description{
Predict the NFL
}
