% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_season.R
\name{simulate_season}
\alias{simulate_season}
\title{Simulation Seasons}
\usage{
simulate_season(year, seed = 123, simulations = 1000, preseason = TRUE)
}
\arguments{
\item{year:}{The season you wish to simulate}

\item{simulations:}{how many times do you wish to simulate the season?}

\item{seed:}{seed for the model. pick a random number.}

\item{preseason:}{this will simulate the season from scratch based on latest performance data that account for offseason regression to the mean.}
}
\value{
dataset of results for season simulations
}
\description{
Simulation Seasons
}
