% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_season.R
\name{simulated_season_table}
\alias{simulated_season_table}
\title{Create a table for simulated Seasons}
\usage{
simulated_season_table(season_simulation_data)
}
\arguments{
\item{season_simulation_data:}{the output of the season simulation function}
}
\value{
makes the table pretty!
}
\description{
Create a table for simulated Seasons
}
